requirements.txt
# Python 3.11+ required
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2

